@@ -1,14 +1,11 @@
-module github.com/blueimp/aws-smtp-relay
+module github.com/KamorionLabs/aws-smtp-relay
 
-go 1.14
+go 1.21
 
 require (
 	github.com/aws/aws-sdk-go v1.38.61
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/go-sql-driver/mysql v1.5.0 // indirect
-	github.com/kr/pretty v0.2.0 // indirect
 	github.com/mhale/smtpd v0.0.0-20210322105601-438c8edb069c
-	github.com/stretchr/testify v1.5.1 // indirect
 	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
+
+require github.com/jmespath/go-jmespath v0.4.0 // indirect
