@@ -2,15 +2,15 @@ package relay
 
 import (
 	"errors"
-	"io/ioutil"
+	"io"
 	"net"
 	"os"
 	"regexp"
 	"testing"
 
+	"github.com/KamorionLabs/aws-smtp-relay/internal/relay"
 	"github.com/aws/aws-sdk-go/service/pinpointemail"
 	"github.com/aws/aws-sdk-go/service/pinpointemail/pinpointemailiface"
-	"github.com/blueimp/aws-smtp-relay/internal/relay"
 )
 
 var testData = struct {
@@ -69,8 +69,8 @@ func sendHelper(
 		outWriter.Close()
 		errWriter.Close()
 	}()
-	stdout, _ := ioutil.ReadAll(outReader)
-	stderr, _ := ioutil.ReadAll(errReader)
+	stdout, _ := io.ReadAll(outReader)
+	stderr, _ := io.ReadAll(errReader)
 	return testData.input, stdout, stderr, sendErr
 }
 
