@@ -2,15 +2,15 @@ package relay
 
 import (
 	"errors"
-	"io/ioutil"
+	"io"
 	"net"
 	"os"
 	"regexp"
 	"testing"
 
+	"github.com/KamorionLabs/aws-smtp-relay/internal/relay"
 	"github.com/aws/aws-sdk-go/service/ses"
 	"github.com/aws/aws-sdk-go/service/ses/sesiface"
-	"github.com/blueimp/aws-smtp-relay/internal/relay"
 )
 
 var testData = struct {
@@ -38,6 +38,7 @@ func sendHelper(
 	configurationSetName *string,
 	allowFromRegExp *regexp.Regexp,
 	denyToRegExp *regexp.Regexp,
+	arns *relay.ARNs,
 	apiErr error,
 ) (email *ses.SendRawEmailInput, out []byte, err []byte, sendErr error) {
 	outReader, outWriter, _ := os.Pipe()
@@ -58,14 +59,15 @@ func sendHelper(
 			setName:         configurationSetName,
 			allowFromRegExp: allowFromRegExp,
 			denyToRegExp:    denyToRegExp,
+			arns:            arns,
 		}
 		testData.err = apiErr
 		sendErr = c.Send(origin, from, to, data)
 		outWriter.Close()
 		errWriter.Close()
 	}()
-	stdout, _ := ioutil.ReadAll(outReader)
-	stderr, _ := ioutil.ReadAll(errReader)
+	stdout, _ := io.ReadAll(outReader)
+	stderr, _ := io.ReadAll(errReader)
 	return testData.input, stdout, stderr, sendErr
 }
 
@@ -75,7 +77,7 @@ func TestSend(t *testing.T) {
 	to := []string{"bob@example.org"}
 	data := []byte{'T', 'E', 'S', 'T'}
 	setName := ""
-	input, out, err, _ := sendHelper(&origin, from, to, data, &setName, nil, nil, nil)
+	input, out, err, _ := sendHelper(&origin, from, to, data, &setName, nil, nil, nil, nil)
 	if *input.Source != from {
 		t.Errorf(
 			"Unexpected source: %s. Expected: %s",
@@ -115,7 +117,7 @@ func TestSendWithMultipleRecipients(t *testing.T) {
 	to := []string{"bob@example.org", "charlie@example.org"}
 	data := []byte{'T', 'E', 'S', 'T'}
 	setName := ""
-	input, out, err, _ := sendHelper(&origin, from, to, data, &setName, nil, nil, nil)
+	input, out, err, _ := sendHelper(&origin, from, to, data, &setName, nil, nil, nil, nil)
 	if len(input.Destinations) != 2 {
 		t.Errorf(
 			"Unexpected number of destinations: %d. Expected: %d",
@@ -145,7 +147,7 @@ func TestSendWithDeniedSender(t *testing.T) {
 	data := []byte{'T', 'E', 'S', 'T'}
 	setName := ""
 	regexp, _ := regexp.Compile(`^admin@example\.org$`)
-	input, out, err, sendErr := sendHelper(&origin, from, to, data, &setName, regexp, nil, nil)
+	input, out, err, sendErr := sendHelper(&origin, from, to, data, &setName, regexp, nil, nil, nil)
 	if input != nil {
 		t.Errorf(
 			"Unexpected number of destinations: %d. Expected: %d",
@@ -171,7 +173,7 @@ func TestSendWithDeniedRecipient(t *testing.T) {
 	data := []byte{'T', 'E', 'S', 'T'}
 	setName := ""
 	regexp, _ := regexp.Compile(`^bob@example\.org$`)
-	input, out, err, sendErr := sendHelper(&origin, from, to, data, &setName, nil, regexp, nil)
+	input, out, err, sendErr := sendHelper(&origin, from, to, data, &setName, nil, regexp, nil, nil)
 	if len(input.Destinations) != 1 {
 		t.Errorf(
 			"Unexpected number of destinations: %d. Expected: %d",
@@ -204,7 +206,7 @@ func TestSendWithApiError(t *testing.T) {
 	data := []byte{'T', 'E', 'S', 'T'}
 	setName := ""
 	apiErr := errors.New("API failure")
-	input, out, err, sendErr := sendHelper(&origin, from, to, data, &setName, nil, nil, apiErr)
+	input, out, err, sendErr := sendHelper(&origin, from, to, data, &setName, nil, nil, nil, apiErr)
 	if *input.Source != from {
 		t.Errorf(
 			"Unexpected source: %s. Expected: %s",
@@ -245,7 +247,15 @@ func TestNew(t *testing.T) {
 	setName := ""
 	allowFromRegExp, _ := regexp.Compile(`^admin@example\.org$`)
 	denyToRegExp, _ := regexp.Compile(`^bob@example\.org$`)
-	client := New(&setName, allowFromRegExp, denyToRegExp)
+	sourceArn := "arn:aws:ses:us-east-1:123456789012:identity/example.com"
+	fromArn := "arn:aws:ses:us-east-1:123456789012:identity/from.example.com"
+	returnPathArn := "arn:aws:ses:us-east-1:123456789012:identity/return.example.com"
+	arns := &relay.ARNs{
+		SourceArn:     &sourceArn,
+		FromArn:       &fromArn,
+		ReturnPathArn: &returnPathArn,
+	}
+	client := New(&setName, allowFromRegExp, denyToRegExp, arns)
 	_, ok := interface{}(client).(relay.Client)
 	if !ok {
 		t.Error("Unexpected: client is not a relay.Client")
@@ -258,5 +268,17 @@ func TestNew(t *testing.T) {
 	}
 	if client.denyToRegExp != denyToRegExp {
 		t.Errorf("Unexpected denyToRegExp: %s", client.denyToRegExp)
+	}
+	if client.arns != arns {
+		t.Errorf("Unexpected arns: %v", client.arns)
+	}
+	if client.arns.SourceArn != &sourceArn {
+		t.Errorf("Unexpected sourceArn: %s", *client.arns.SourceArn)
+	}
+	if client.arns.FromArn != &fromArn {
+		t.Errorf("Unexpected fromArn: %s", *client.arns.FromArn)
+	}
+	if client.arns.ReturnPathArn != &returnPathArn {
+		t.Errorf("Unexpected returnPathArn: %s", *client.arns.ReturnPathArn)
 	}
 }
