@@ -2,13 +2,12 @@ package main
 
 import (
 	"flag"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"testing"
 
-	pinpointrelay "github.com/blueimp/aws-smtp-relay/internal/relay/pinpoint"
-	sesrelay "github.com/blueimp/aws-smtp-relay/internal/relay/ses"
+	pinpointrelay "github.com/KamorionLabs/aws-smtp-relay/internal/relay/pinpoint"
+	sesrelay "github.com/KamorionLabs/aws-smtp-relay/internal/relay/ses"
 )
 
 const certPEM = `-----BEGIN CERTIFICATE-----
@@ -95,7 +94,7 @@ uEGsi+l2fTj/F+eZLE6sYoMprgJrbfeqtRWFguUgTn7s5hfU0tZ46al5d0vz8fWK
 var sampleHash = "$2y$10$85/eICRuwBwutrou64G5HeoF3Ek/qf1YKPLba7ckiMxUTAeLIeyaC"
 
 func createTmpFile(content string) (fileName *string, err error) {
-	file, err := ioutil.TempFile("", "")
+	file, err := os.CreateTemp("", "")
 	if err != nil {
 		return
 	}
