@@ -4,10 +4,10 @@ import (
 	"net"
 	"regexp"
 
+	"github.com/KamorionLabs/aws-smtp-relay/internal/relay"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ses"
 	"github.com/aws/aws-sdk-go/service/ses/sesiface"
-	"github.com/blueimp/aws-smtp-relay/internal/relay"
 )
 
 // Client implements the Relay interface.
@@ -16,6 +16,7 @@ type Client struct {
 	setName         *string
 	allowFromRegExp *regexp.Regexp
 	denyToRegExp    *regexp.Regexp
+	arns            *relay.ARNs
 }
 
 // Send uses the client SESAPI to send email data
@@ -35,12 +36,18 @@ func (c Client) Send(
 		relay.Log(origin, &from, deniedRecipients, err)
 	}
 	if len(allowedRecipients) > 0 {
-		_, err := c.sesAPI.SendRawEmail(&ses.SendRawEmailInput{
+		input := &ses.SendRawEmailInput{
 			ConfigurationSetName: c.setName,
 			Source:               &from,
 			Destinations:         allowedRecipients,
 			RawMessage:           &ses.RawMessage{Data: data},
-		})
+		}
+		if c.arns != nil {
+			input.SourceArn = c.arns.SourceArn
+			input.FromArn = c.arns.FromArn
+			input.ReturnPathArn = c.arns.ReturnPathArn
+		}
+		_, err := c.sesAPI.SendRawEmail(input)
 		relay.Log(origin, &from, allowedRecipients, err)
 		if err != nil {
 			return err
@@ -54,11 +61,13 @@ func New(
 	configurationSetName *string,
 	allowFromRegExp *regexp.Regexp,
 	denyToRegExp *regexp.Regexp,
+	arns *relay.ARNs,
 ) Client {
 	return Client{
 		sesAPI:          ses.New(session.Must(session.NewSession())),
 		setName:         configurationSetName,
 		allowFromRegExp: allowFromRegExp,
 		denyToRegExp:    denyToRegExp,
+		arns:            arns,
 	}
 }
